+++ internal/members/evaluations.go
@@ -13,8 +13,8 @@ type Evaluation struct {
 
 // MemberEvaluation represents an Evaluation belonging to a Member
 type MemberEvaluation struct {
-	ID             int64                `json:"id" bson:"id"`
-	MemberID       int64                `json:"memberId" bson:"memberId"`
+	ID             int                `json:"id" bson:"id"`
+	MemberID       int                `json:"memberId" bson:"memberId"`
 	Name           string               `json:"name" bson:"name"`
 	StartDate      string               `json:"startDate" bson:"startDate"`
 	EndDate        string               `json:"endDate" bson:"endDate"`
@@ -34,7 +34,7 @@ type EvaluationActivity struct {
 
 // EvaluationsByMemberID fetches all evaluation records for a member
 // Received a member id, return a []MemberEvaluation
-func EvaluationsByMemberID(id int64) ([]MemberEvaluation, error) {
+func EvaluationsByMemberID(id int) ([]MemberEvaluation, error) {
 
 	es := []MemberEvaluation{}
